+++ src/main/java/com/lch/topick/web/myLocationSet/entity/MyLocationSet.java
@@ -21,7 +21,7 @@ public class MyLocationSet {
 	
 	@Id	//주식별자	
 	@GeneratedValue(strategy=GenerationType.IDENTITY )	//Auto_Increment
-	private int addressNo;			
+	private long addressNo;			
 	
 	private String memberId;		//사용자 고유ID (member table 참조)
 	private String addressPostcode;	//우편번호, API: zonecode
